Add files via upload
BitLocker Recovery for live Windows system logged in with Admin rights.

Place the script on an external drive.  On target system, open terminal in admin, cd to the external drive containing the script, execute by typing "Get-BitLockerInfo.ps1".  Script will recovery BitLocker encryption keys for all volumes on the system and kick the data out to a text file on the drive where script was ran from.  If you get a security error, type "Set-ExecutionPolicy RemoteSigned" and then run the script again.